+++ src/Sentry.Unity/Integrations/UnityScopeIntegration.cs
@@ -34,11 +34,13 @@ internal class UnityScopeUpdater
 {
     private readonly SentryUnityOptions _options;
     private readonly IApplication _application;
+    private readonly ISceneManager _sceneManager;
 
-    public UnityScopeUpdater(SentryUnityOptions options, IApplication application)
+    public UnityScopeUpdater(SentryUnityOptions options, IApplication application, ISceneManager? sceneManager = null)
     {
         _options = options;
         _application = application;
+        _sceneManager = sceneManager ?? SceneManagerAdapter.Instance;
     }
 
     public void ConfigureScope(Scope scope)
@@ -139,6 +141,7 @@ internal class UnityScopeUpdater
         unity.TargetFrameRate = MainThreadData.TargetFrameRate;
         unity.CopyTextureSupport = MainThreadData.CopyTextureSupport;
         unity.RenderingThreadingMode = MainThreadData.RenderingThreadingMode;
+        unity.ActiveSceneName = _sceneManager.GetActiveScene().Name;
     }
 
     private void PopulateTags(Action<string, string> setTag)